# apps/agent-ui/src/lib/agentApi.ts
const API_URL = import.meta.env.VITE_AGENT_API_URL || 'http://localhost:3005';

export interface WorkflowDefinition {
  name: string;
  nodes: Array<Record<string, unknown>>;
  connections: Record<string, unknown>;
  settings?: Record<string, unknown>;
  meta?: Record<string, unknown>;
}

export interface WorkflowBlueprint {
  intentSummary: string;
  triggers: Array<{ type: 'webhook' | 'scheduleTrigger'; config: Record<string, unknown> }>;
  logic: Array<{ type: 'if' | 'splitInBatches'; config: Record<string, unknown> }>;
  executors: Array<{ type: 'set' | 'httpRequest'; config: Record<string, unknown> }>;
  missingFields: string[];
}

export interface InteractionOption {
  label: string;
  value: string;
}

export interface InteractionRequest {
  id: string;
  mode: 'single' | 'multi' | 'image';
  field:
    | 'tts_voice'
    | 'screen_emoji'
    | 'chassis_action'
    | 'hand_gestures'
    | 'yolo_gestures'
    | 'emotion_labels'
    | 'arm_actions'
    | 'face_profiles';
  title: string;
  description?: string;
  options: InteractionOption[];
  minSelections?: number;
  maxSelections?: number;
  selected?: string | string[];
  allowUpload?: boolean;
  uploadHint?: string;
}

export interface AgentResponse {
  type:
    | 'guidance'
    | 'summary_ready'
    | 'workflow_ready'
    | 'error'
    | 'select_single'
    | 'select_multi'
    | 'image_upload';
  message: string;
  blueprint?: WorkflowBlueprint;
  confirmedEntities?: Record<string, string>;
  missingInfo?: string[];
  workflow?: WorkflowDefinition;
  reasoning?: string;
  interaction?: InteractionRequest;
  metadata?: {
    iterations?: number;
    nodeCount?: number;
    showContinueButton?: boolean;
    showConfirmBuildButton?: boolean;
  };
  details?: unknown;
}

export interface AgentChatResponse {
  sessionId: string;
  response: AgentResponse;
}

export interface WorkflowCreateResult {
  workflowId: string;
  workflowName?: string;
  workflowUrl?: string;
}

export interface UploadFaceImageResult {
  success: boolean;
  profile?: string;
  fileId?: string;
  fileName?: string;
  url?: string;
}

async function postJson<T>(path: string, body: unknown): Promise<T> {
  const response = await fetch(`${API_URL}${path}`, {
    method: 'POST',
    headers: { 'Content-Type': 'application/json' },
    body: JSON.stringify(body),
  });

  if (!response.ok) {
    const error = await response.json().catch(() => ({}));
    throw new Error(error.error || 'Request failed');
  }

  return response.json() as Promise<T>;
}

export async function uploadFaceImage(
  profile: string,
  fileName: string,
  contentBase64: string
): Promise<UploadFaceImageResult> {
  const response = await fetch(`${API_URL}/api/agent/upload-face`, {
    method: 'POST',
    headers: { 'Content-Type': 'application/json' },
    body: JSON.stringify({ profile, fileName, contentBase64 }),
  });

  if (!response.ok) {
    const error = await response.json().catch(() => ({}));
    throw new Error(error.error || 'Upload failed');
  }

  return response.json() as Promise<UploadFaceImageResult>;
}

export async function sendAgentMessage(message: string, sessionId?: string): Promise<AgentChatResponse> {
  return postJson<AgentChatResponse>('/api/agent/chat', { message, sessionId });
}

export async function confirmAgentWorkflow(sessionId: string): Promise<AgentChatResponse> {
  return postJson<AgentChatResponse>('/api/agent/confirm-build', { sessionId });
}

export async function resetAgentSession(sessionId: string): Promise<{ success: boolean }> {
  return postJson<{ success: boolean }>('/api/agent/reset-session', { sessionId });
}

export async function createWorkflow(
  workflow: WorkflowDefinition,
  sessionId?: string
): Promise<WorkflowCreateResult> {
  return postJson<WorkflowCreateResult>('/api/workflow/create', {
    workflow,
    sessionId,
  });
}