# docs/api/agent-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "n8n-mcp Agent Backend API",
    "version": "1.0.0",
    "description": "Agent backend HTTP API for local frontend development."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3005"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" }
                  },
                  "required": ["status"]
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/chat": {
      "post": {
        "summary": "Chat with agent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponseEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/agent/confirm": {
      "post": {
        "summary": "Confirm summary / advance phase",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfirmRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgentResponseEnvelope" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/agent/confirm-build": {
      "post": {
        "summary": "Confirm build (alias of confirm)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfirmRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgentResponseEnvelope" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/agent/reset-session": {
      "post": {
        "summary": "Reset session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfirmRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  },
                  "required": ["success"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/agent/upload-face": {
      "post": {
        "summary": "Upload face image (base64)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UploadFaceRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UploadFaceResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/workflow/create": {
      "post": {
        "summary": "Create workflow in n8n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WorkflowCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created workflow",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorkflowCreateResult" }
              }
            }
          },
          "400": {
            "description": "Bad request or missing n8n config",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "sessionId": { "type": "string" }
        },
        "required": ["message"]
      },
      "ConfirmRequest": {
        "type": "object",
        "properties": {
          "sessionId": { "type": "string" }
        },
        "required": ["sessionId"]
      },
      "AgentResponseEnvelope": {
        "type": "object",
        "properties": {
          "sessionId": { "type": "string" },
          "response": { "$ref": "#/components/schemas/AgentResponse" }
        },
        "required": ["sessionId", "response"]
      },
      "AgentResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "guidance",
              "summary_ready",
              "workflow_ready",
              "error",
              "select_single",
              "select_multi",
              "image_upload"
            ]
          },
          "message": { "type": "string" },
          "blueprint": { "$ref": "#/components/schemas/WorkflowBlueprint" },
          "workflow": { "$ref": "#/components/schemas/WorkflowDefinition" },
          "reasoning": { "type": "string" },
          "interaction": { "$ref": "#/components/schemas/InteractionRequest" },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["type", "message"]
      },
      "InteractionRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "mode": { "type": "string", "enum": ["single", "multi", "image"] },
          "field": {
            "type": "string",
            "enum": [
              "tts_voice",
              "screen_emoji",
              "chassis_action",
              "hand_gestures",
              "yolo_gestures",
              "emotion_labels",
              "arm_actions",
              "face_profiles"
            ]
          },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": { "type": "string" },
                "value": { "type": "string" }
              },
              "required": ["label", "value"]
            }
          },
          "minSelections": { "type": "integer" },
          "maxSelections": { "type": "integer" },
          "selected": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "array",
                "items": { "type": "string" }
              }
            ]
          },
          "allowUpload": { "type": "boolean" },
          "uploadHint": { "type": "string" }
        },
        "required": ["id", "mode", "field", "title", "options"]
      },
      "WorkflowBlueprint": {
        "type": "object",
        "properties": {
          "intentSummary": { "type": "string" },
          "triggers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BlueprintEntry" }
          },
          "logic": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BlueprintEntry" }
          },
          "executors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BlueprintEntry" }
          },
          "missingFields": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "BlueprintEntry": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "config": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "WorkflowDefinition": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "nodes": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "connections": {
            "type": "object",
            "additionalProperties": true
          },
          "settings": {
            "type": "object",
            "additionalProperties": true
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["name", "nodes", "connections"]
      },
      "WorkflowCreateRequest": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "workflow": { "$ref": "#/components/schemas/WorkflowDefinition" }
            },
            "required": ["workflow"]
          },
          {
            "type": "object",
            "properties": {
              "sessionId": { "type": "string" }
            },
            "required": ["sessionId"]
          }
        ]
      },
      "UploadFaceRequest": {
        "type": "object",
        "properties": {
          "profile": { "type": "string" },
          "fileName": { "type": "string" },
          "contentBase64": { "type": "string" }
        },
        "required": ["profile", "fileName", "contentBase64"]
      },
      "UploadFaceResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "profile": { "type": "string" },
          "fileId": { "type": "string" },
          "fileName": { "type": "string" },
          "url": { "type": "string" }
        },
        "required": ["success", "fileName", "url"]
      },
      "WorkflowCreateResult": {
        "type": "object",
        "properties": {
          "workflowId": { "type": "string" },
          "workflowName": { "type": "string" },
          "workflowUrl": { "type": "string" }
        },
        "required": ["workflowId", "workflowName", "workflowUrl"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      }
    }
  }
}